+++ etna/models/sklearn.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import List
 from typing import Optional
 
@@ -17,6 +18,41 @@ class _SklearnAdapter(BaseAdapter):
         self.model = regressor
         self.regressor_columns: Optional[List[str]] = None
 
+    def _check_not_used_columns(self, df: pd.DataFrame):
+        if self.regressor_columns is None:
+            raise ValueError("Something went wrong, regressor_columns is None!")
+
+        columns_not_used = [col for col in df.columns if col not in ["target", "timestamp"] + self.regressor_columns]
+        if columns_not_used:
+            warnings.warn(
+                message=f"This model doesn't work with exogenous features unknown in future. "
+                f"Columns {columns_not_used} won't be used."
+            )
+
+    def _select_regressors(self, df: pd.DataFrame) -> Optional[pd.DataFrame]:
+        """Select data with regressors.
+
+        During fit there can't be regressors with NaNs, they are removed at higher level.
+        Look at the issue: https://github.com/tinkoff-ai/etna/issues/557
+
+        During prediction without validation NaNs in regressors can lead to exception from the underlying model,
+        but it depends on the model, so it was decided to not validate this.
+
+        This model requires data to be in numeric dtype.
+        """
+        if self.regressor_columns is None:
+            raise ValueError("Something went wrong, regressor_columns is None!")
+
+        if self.regressor_columns:
+            try:
+                result = df[self.regressor_columns].apply(pd.to_numeric)
+            except ValueError as e:
+                raise ValueError(f"Only convertible to numeric features are allowed! Error: {str(e)}")
+        else:
+            raise ValueError("There are not features for fitting the model!")
+
+        return result
+
     def fit(self, df: pd.DataFrame, regressors: List[str]) -> "_SklearnAdapter":
         """
         Fit Sklearn model.
@@ -34,10 +70,8 @@ class _SklearnAdapter(BaseAdapter):
             Fitted model
         """
         self.regressor_columns = regressors
-        try:
-            features = df[self.regressor_columns].apply(pd.to_numeric)
-        except ValueError:
-            raise ValueError("Only convertible to numeric features are accepted!")
+        self._check_not_used_columns(df)
+        features = self._select_regressors(df)
         target = df["target"]
         self.model.fit(features, target)
         return self
@@ -56,10 +90,7 @@ class _SklearnAdapter(BaseAdapter):
         :
             Array with predictions
         """
-        try:
-            features = df[self.regressor_columns].apply(pd.to_numeric)
-        except ValueError:
-            raise ValueError("Only convertible to numeric features are accepted!")
+        features = self._select_regressors(df)
         pred = self.model.predict(features)
         return pred